backend/requirements.txt
fastapi
pandas
SQLAlchemy
python-dotenv
openai
uvicorn
python-multipart
langchain
langchain-community
langchain-openai